feature: fix bug when user revise password, fix bug when user delete account, change way to delete object in canvas at meme generator page

+++ src/components/SettingPage/DeleteAccount.js
@@ -2,7 +2,7 @@ import React, { useRef } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
 
-import { alertDelete } from '../../utlis/alert';
+import { alertDelete, alertError } from '../../utlis/alert';
 import color from '../Styled/colorTheme';
 import { deleteAllData, deleteAccount, reAuth } from '../../utlis/firebase';
 
@@ -49,28 +49,33 @@ const Button0 = styled.div`
 `;
 
 function DeleteAccount() {
-    const userData = useSelector((state) => state.userData);
-    const password = useRef(null);
+  const userData = useSelector((state) => state.userData);
+  const password = useRef(null);
 
-    const deleteTheUser = () => {
-        reAuth(password.current.value)
+  const deleteTheUser = () => {
+    reAuth(password.current.value)
+      .then((res) => {
+        if (!res) { 
+          alertError(undefined, '密碼輸入錯誤！請重新輸入！'); 
+        }
+        if (res) {
+          deleteAllData(userData.user_id)
             .then(() => {
-                deleteAllData(userData.user_id)
-                    .then(() => {
-                        deleteAccount(password.current.value);
-                    })
+              deleteAccount(password.current.value);
             })
-    }
+        }
+      })
+  }
 
-    return (
-        <Container6>
-            <Container7>刪除帳戶</Container7>
-            <p>請輸入密碼：</p>
-            <Input1 type="password" ref={password} />
-            <br></br>
-            <Button0 color={color} onClick={() => alertDelete(deleteTheUser)}>刪除帳戶</Button0>
-        </Container6>
-    )
+  return (
+    <Container6>
+      <Container7>刪除帳戶</Container7>
+      <p>請輸入密碼：</p>
+      <Input1 type="password" ref={password} />
+      <br></br>
+      <Button0 color={color} onClick={() => alertDelete(deleteTheUser)}>刪除帳戶</Button0>
+    </Container6>
+  )
 }
 
 export default DeleteAccount;
